# openapi.yaml
openapi: 3.1.0

info:
  title: Avito Stats API
  description: API that allows to track amount of ads grouped by search query and region
  version: 1.0.0

servers:
  - url: http://localhost
    description: local server for development

paths:
  /add:
    post:
      summary: adds pair of query and location_id to tracking schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddScheme'
      responses:
        '200':
          description: Returns id of created pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid  # todo update
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /stop/{pair_id}:
    parameters:
      - in: path
        name: pair_id
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Stop tracking pair
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'

  /stat:
    post:
      summary: Endpoint for getting number of ads between specified datetime_from, datetime_to (all in UTC)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatsSchemaRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsSchemaResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /top:
    post:
      summary: Endpoint similar to /stat but returns top 5 ads for every hour between specified datetime_from and datetime_to.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatsSchemeRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopAdsResponse'

        '404':
          $ref: '#/components/responses/NotFound'

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /location:
    description: Get avito location ID by name (region name or city)
    post:
      description: Return location id and specified in request payload location name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                location_name: string
        
      responses:
        '200':
          description: 'OK'
          content:
            application/json:
              schema:
                type: object
                properties:
                  location_id: 
                    type: int
                  location_name:
                    type: string

        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'


components:
  responses:
    NotFound:
      description: The specified pair_id was not found.
      content:
        application/json:
          schema:
            $ref: '#/schemas/Error'

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
      additionalProperties: false

    StatItem:
      type: object
      properties:
        moment:
          type: string
          format: date-time
        count:
          type: integer

    AdItem:
      type: object
      required:
        - ad_id
        - url
        - moment
      properties:
        ad_id:
          type: integer
        url:
          type: string
        moment:
          type: string
          format: date-time

    AddScheme:
      type: object
      required:
        - query
        - location_id
      properties:
        query:
          type: string
        location_id:
          type: int
        check_every:
          type: int
          default: 60

    StatsSchemaRequest:
      type: object
      required:
        - id
        - datetime_from
        - datetime_to
      properties:
        id:
          type: string
          format: uuid
        datetime_from:
          type: string
          format: date-time
        datetime_to:
          type: string
          format: date-time

    StatsSchemaResponse:
      type: object
      required:
        - query
        - region
        - stats
      properties:
        id:
          type: string
          format: uuid
        stats:
          type: array
          items:
            $ref: '#/StatItem'

    TopAdsResponse:
      type: object
      properties:
        query:
          type: string
        region:
          type: string
        top_ads:
          type: array
          items:
            $ref: '#/AdItem'
             
    
         
         